Prepare mixins
